requirements.txt
fastapi~=0.116.1
python-dotenv~=1.1.1
pydantic~=2.11.7
requests~=2.32.4
python-docx~=1.2.0
beautifulsoup4~=4.13.4
langchain-text-splitters~=0.3.9
numpy~=2.2.6
langchain-core~=0.3.72
langsmith~=0.4.11
sentence-transformers~=5.0.0
protobuf~=6.31.1
langchain-google-genai~=2.1.9
uvicorn
pdfplumber
faiss-cpu==1.7.4